Enh: Improve memories

+++ app/api/cloudflare/delete-videos/route.ts
@@ -1,18 +1,28 @@
+import { createClient } from '@/utils/supabase/server'
 import { NextResponse } from 'next/server'
 
 export async function GET() {
-  const endpoint = `https://api.cloudflare.com/client/v4/accounts/${process.env.CLOUDFLARE_ACCOUNT_ID}/stream`
+  const cloudflareStreamEndpoint = `https://api.cloudflare.com/client/v4/accounts/${process.env.CLOUDFLARE_ACCOUNT_ID}/stream`
+  const supabase = await createClient()
 
-  const response = await fetch(endpoint, {
+  const { data: memories, error } = await supabase.from('memories').select('cloudflare_file_id')
+
+  if (error) {
+    console.error('Error fetching memories:', error)
+    return NextResponse.json({ error: 'Failed to fetch memories' }, { status: 500 })
+  }
+
+  const listVideosData = await fetch(cloudflareStreamEndpoint, {
     method: 'GET',
     headers: {
       Authorization: `bearer ${process.env.CLOUDFLARE_API_TOKEN}`,
     },
-  })
+  }).then(res => res.json())
 
-  const data = await response.json()
-  data.result.forEach(async (item: any) => {
-    const response = await fetch(`${endpoint}/${item.uid}`, {
+  listVideosData.result.forEach(async (item: any) => {
+    if (memories.some(memory => memory.cloudflare_file_id === item.uid)) return
+
+    const response = await fetch(`${cloudflareStreamEndpoint}/${item.uid}`, {
       method: 'DELETE',
       headers: {
         Authorization: `bearer ${process.env.CLOUDFLARE_API_TOKEN}`,
